Rename Java as well
