+++ sdk/src/wallet/bindings/java/examples/src/BurnNft.java
@@ -22,7 +22,7 @@ public class BurnNft {
         );
 
         // Get account and sync it with the registered node to ensure that its balances are up-to-date.
-        AccountHandle a = wallet.getAccount(new AccountAlias(Env.ACCOUNT_NAME));
+        Account a = wallet.getAccount(new AccountAlias(Env.ACCOUNT_NAME));
         a.syncAccount(new SyncAccount().withOptions(new SyncOptions()));
 
         // TODO: replace with your own values.
