+++ sdk/src/wallet/bindings/java/examples/src/SendAmount.java
@@ -24,7 +24,7 @@ public class SendAmount {
         );
 
         // Get account and sync it with the registered node to ensure that its balances are up-to-date.
-        AccountHandle account = wallet.getAccount(Env.ACCOUNT_NAME);
+        Account account = wallet.getAccount(Env.ACCOUNT_NAME);
 
         AddressWithAmount[] outputs = new AddressWithAmount[] {
                         new AddressWithAmount()
