+++ sdk/src/wallet/bindings/java/lib/src/test/java/org/iota/AccountTests.java
@@ -3,7 +3,7 @@
 package org.iota;
 
 import org.iota.types.AccountAddress;
-import org.iota.types.AccountHandle;
+import org.iota.types.Account;
 import org.iota.types.account_methods.SetAlias;
 import org.iota.types.exceptions.WalletException;
 import org.iota.types.ids.account.AccountIndex;
@@ -13,7 +13,7 @@ public class AccountTests extends TestSettings {
 
     @Test
     public void testListAddresses() throws WalletException {
-        AccountHandle a = wallet.createAccount("Alice");
+        Account a = wallet.createAccount("Alice");
         AccountAddress[] addresses = a.getAddresses();
         for(AccountAddress address : addresses) {
             System.out.println(address.getAddress());
@@ -22,7 +22,7 @@ public class AccountTests extends TestSettings {
 
     @Test
     public void testSetAlias() throws WalletException {
-        AccountHandle a = wallet.createAccount("Alice");
+        Account a = wallet.createAccount("Alice");
         a.setAlias(new SetAlias().withAlias("Bob"));
         System.out.println(wallet.getAccount(new AccountIndex(0)).getAlias());
     }
